ExchangeResumeScreen pt1
