+++ src/routes/index.tsx
@@ -37,8 +37,8 @@ const Stack = createNativeStackNavigator<RootStackParamList>();
 
 const Routes: React.FC = () => {
 	const walletState = useState(globalWalletState());
-	const initialScreen = walletState.value.walletId ? 'WalletScreen' : 'WelcomeScreen';
-	// const initialScreen = 'ExchangeScreen';
+	// const initialScreen = walletState.value.walletId ? 'WalletScreen' : 'WelcomeScreen';
+	const initialScreen = 'ExchangeResumeScreen';
 
 	return (
 		<NavigationContainer>
